+++ shared/constants/reducer.js
@@ -13,7 +13,7 @@ import type {State as Notification} from '../constants/types/notifications'
 import type {State as Pgp} from '../constants/types/pgp'
 import type {State as Pinentry} from '../constants/types/pinentry'
 import type {State as Profile} from '../constants/types/profile'
-import type {State as Push} from '../constants/push'
+import type {State as Push} from '../constants/types/push'
 import type {State as RouteTree} from '../constants/route-tree'
 import type {State as Settings} from '../constants/settings'
 import type {State as Signup} from '../constants/signup'
